--- opentelekomcloud/acceptance/antiddos/import_opentelekomcloud_antiddos_v1_test.go
@@ -1,30 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccAntiDdosV1_importBasic(t *testing.T) {
-	resourceName := "opentelekomcloud_antiddos_v1.antiddos_1"
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCheckAntiDdosV1Destroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccAntiDdosV1_basic,
-			},
-
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
