+++ opentelekomcloud/acceptance/antiddos/utils.go
@@ -1,0 +1,3 @@
+package acceptance
+
+var testServiceV1 = "antiddos"
